refactor(be): extract inline Pydantic schemas to domain schemas.py modules (#252)
Move all inline BaseModel classes from 10 routers to their respective
domain/x/schemas.py files. Pure refactor — no field renames, no
behaviour change. Creates users, workspaces, storage, lots, tags, and
custom_fields schemas.py; extends parts and projects schemas.py.
